[BE #99] refactor : ReservationInterceptor 수정
생성자 주입으로 수정
root 페이지 주소 yml에 추가
