+++ BE/src/test/resources/application.yml
@@ -14,6 +14,7 @@ logging:
     kr.codesquad.airbnb11: debug
 
 host: localhost:8080
+root: localhost:8080/
 github:
   client-id: 9875df2fa499cdb9ff25
   client-secret: d1dff8a224738636f3fb0e6a3781b5ee8c17f494
